pyproject.toml
[project]
name = "google-adk-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-adk>=1.11.0",
    "exa-py>=1.0.7",
    "psycopg2-binary>=2.9.10",
    "agentops>=0.4.20",
]
